+++ aggregator/src/aggregation/config.rs
@@ -1,14 +1,9 @@
 use halo2_proofs::{
-    halo2curves::bn256::{Fq, Fr, G1Affine},
-    plonk::{Column, ConstraintSystem, Instance},
+    halo2curves::bn256::Fr,
+    plonk::{Circuit, Column, ConstraintSystem, Instance},
 };
-use snark_verifier::{
-    loader::halo2::halo2_ecc::{
-        ecc::{BaseFieldEccChip, EccChip},
-        fields::fp::FpConfig,
-        halo2_base::gates::{flex_gate::FlexGateConfig, range::RangeConfig},
-    },
-    util::arithmetic::modulus,
+use snark_verifier::loader::halo2::halo2_ecc::halo2_base::gates::circuit::{
+    builder::BaseCircuitBuilder, BaseConfig,
 };
 use zkevm_circuits::{
     keccak_circuit::{KeccakCircuitConfig, KeccakCircuitConfigArgs},
@@ -16,11 +11,7 @@ use zkevm_circuits::{
     util::{Challenges, SubCircuitConfig},
 };
 
-use crate::{
-    constants::{BITS, LIMBS},
-    param::ConfigParams,
-    RlcConfig,
-};
+use crate::{ConfigParams, RlcConfig};
 
 #[derive(Debug, Clone)]
 #[rustfmt::skip]
@@ -28,33 +19,25 @@ use crate::{
 /// This config is hard coded for BN256 curve.
 pub struct AggregationConfig {
     /// Non-native field chip configurations
-    pub base_field_config: FpConfig<Fr, Fq>,
+    pub base_field_config: BaseConfig<Fr>,
     /// Keccak circuit configurations
     pub keccak_circuit_config: KeccakCircuitConfig<Fr>,    
     /// RLC config
     pub rlc_config: RlcConfig,
-    /// Instance for public input; stores
-    /// - accumulator from aggregation (12 elements)
-    /// - batch_public_input_hash (32 elements)
-    /// - the number of valid SNARKs (1 element)
-    pub instance: Column<Instance>,
+    /// There are two instances for public input;
+    /// - accumulator from aggregation, 12 elements, stored in base_field_config.instance
+    /// - batch_public_input_hash, 32 elements, stored here
+    /// - the number of valid SNARKs, 1 element, stored here
+    pub pi_instance: Column<Instance>,
 }
 
 impl AggregationConfig {
-    /// Build a configuration from parameters.
-    pub fn configure(
+    pub(crate) fn configure(
         meta: &mut ConstraintSystem<Fr>,
-        params: &ConfigParams,
+        params: ConfigParams,
         challenges: Challenges,
     ) -> Self {
-        assert!(
-            params.limb_bits == BITS && params.num_limbs == LIMBS,
-            "For now we fix limb_bits = {BITS}, otherwise change code",
-        );
-
-        // RLC configuration
-        let rlc_config = RlcConfig::configure(meta, challenges);
-
+        // It is wired that if we config ...
         // hash configuration for aggregation circuit
         let keccak_circuit_config = {
             let keccak_table = KeccakTable::construct(meta);
@@ -67,21 +50,6 @@ impl AggregationConfig {
 
             KeccakCircuitConfig::new(meta, keccak_circuit_config_args)
         };
-
-        // base field configuration for aggregation circuit
-        let base_field_config = FpConfig::configure(
-            meta,
-            params.strategy.clone(),
-            &params.num_advice,
-            &params.num_lookup_advice,
-            params.num_fixed,
-            params.lookup_bits,
-            BITS,
-            LIMBS,
-            modulus::<Fq>(),
-            0,
-            params.degree as usize,
-        );
 
         let columns = keccak_circuit_config.cell_manager.columns();
         log::info!("keccak uses {} columns", columns.len(),);
@@ -97,38 +65,33 @@ impl AggregationConfig {
         // enable equality for the is_final column
         meta.enable_equality(keccak_circuit_config.keccak_table.is_final);
 
+        let base_field_config = BaseCircuitBuilder::configure_with_params(meta, (&params).into());
+
+        // RLC configuration
+        let rlc_config = RlcConfig::configure(meta, challenges);
+
         // Instance column stores public input column
-        // - the accumulator
         // - the batch public input hash
         // - the number of valid SNARKs
-        let instance = meta.instance_column();
-        meta.enable_equality(instance);
+        let pi_instance = meta.instance_column();
+        meta.enable_equality(pi_instance);
 
         Self {
             base_field_config,
-            rlc_config,
             keccak_circuit_config,
-            instance,
+            rlc_config,
+            pi_instance,
         }
     }
 
-    /// Expose the instance column
-    pub fn instance_column(&self) -> Column<Instance> {
-        self.instance
-    }
-
-    /// Range gate configuration
-    pub fn range(&self) -> &RangeConfig<Fr> {
-        &self.base_field_config.range
-    }
-
-    /// Flex gate configuration
-    pub fn flex_gate(&self) -> &FlexGateConfig<Fr> {
-        &self.base_field_config.range.gate
-    }
-
-    /// Ecc gate configuration
-    pub fn ecc_chip(&self) -> BaseFieldEccChip<G1Affine> {
-        EccChip::construct(self.base_field_config.clone())
+    /// Expose the instance columns
+    /// There are two instances for public input;
+    /// - accumulator from aggregation, 12 elements, stored in base_field_config.instance
+    /// - batch_public_input_hash, 32 elements, stored here
+    /// - the number of valid SNARKs, 1 element, stored here
+    pub fn instance_columns(&self) -> Vec<Column<Instance>> {
+        let mut instances = self.base_field_config.instance.clone();
+        instances.extend([self.pi_instance]);
+        instances
     }
 }
