+++ aggregator/src/lib.rs
@@ -6,13 +6,10 @@ mod batch;
 // This module implements `Chunk` related data types.
 // A chunk is a list of blocks.
 mod chunk;
-// /// proof compression
-// mod compression;
-mod compression_v2;
+/// proof compression
+mod compression;
 /// Configurations
 mod constants;
-/// Core module for circuit assignment
-mod core;
 /// Parameters for compression circuit
 mod param;
 /// utilities
@@ -25,22 +22,12 @@ mod tests;
 pub use aggregation::*;
 pub use batch::BatchHash;
 pub use chunk::ChunkHash;
-// pub use compression::*;
-pub use constants::MAX_AGG_SNARKS;
-
+pub use compression::*;
+pub use constants::{MAX_AGG_SNARKS, *};
 pub use param::*;
 
-pub use compression_v2::*;
-pub use constants::*;
-
 use halo2_proofs::halo2curves::bn256::Bn256;
-use snark_verifier::{
-    pcs::kzg::{Bdfg21, KzgAs, LimbsEncoding},
-    verifier,
-};
+use snark_verifier::pcs::kzg::{Bdfg21, KzgAs};
 
 /// accumulation scheme used for aggregator.
 type AccScheme = KzgAs<Bn256, Bdfg21>;
-/// prove the accumulators are generated faithfully; defer the accumulator check to decider.
-type PlonkSuccinctVerifier =
-    verifier::plonk::PlonkSuccinctVerifier<AccScheme, LimbsEncoding<LIMBS, BITS>>;
